# specs/003-agent-chat/contracts/agents.openapi.yaml
openapi: 3.1.0
info:
  title: Agent Chat API
  version: 0.3.0
  description: Contract for authenticated owner-only agent chat with AI SDK UI message stream responses.
servers:
  - url: http://localhost:3000
security:
  - bearerJwt: []
paths:
  /agents:
    post:
      summary: Create an agent for the authenticated user
      operationId: createAgent
      security:
        - bearerJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedAgent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      summary: List agents owned by the authenticated user
      operationId: listAgents
      security:
        - bearerJwt: []
      responses:
        '200':
          description: Lightweight agents owned by the authenticated user. Role content is intentionally omitted.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /agents/{id}:
    get:
      summary: Retrieve one agent owned by the authenticated user
      operationId: getAgent
      security:
        - bearerJwt: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent details including role instructions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update one agent owned by the authenticated user
      operationId: updateAgent
      security:
        - bearerJwt: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Updated agent details including role instructions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete one agent owned by the authenticated user
      operationId: deleteAgent
      security:
        - bearerJwt: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '204':
          description: Agent deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /agents/{id}/chat:
    post:
      x-speckit-task: T009
      summary: Send a chat request to one owned agent and stream the generated response
      operationId: chatWithAgent
      security:
        - bearerJwt: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentChatRequest'
            examples:
              singleMessage:
                summary: Single user message
                value:
                  input: Hello, what can you help me do?
              conversation:
                summary: Conversation with developer guidance
                value:
                  input:
                    - role: developer
                      content: Keep answers concise.
                    - role: user
                      content: Summarize this project.
      responses:
        '200':
          description: AI SDK UI message stream generated by the agent for @ai-sdk/react useChat.
          headers:
            x-vercel-ai-ui-message-stream:
              schema:
                type: string
                const: v1
              description: Marks the response as an AI SDK UI message stream.
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/ResponseGenerationFailed'
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AgentId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Unique identifier of the agent.
  schemas:
    CreateAgentRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          description: Agent name. Whitespace-only values are invalid.
    UpdateAgentRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          description: New agent name. Whitespace-only values are invalid.
        role:
          type: string
          description: New role instructions. Empty string is allowed.
    AgentChatRequest:
      type: object
      additionalProperties: false
      required:
        - input
      properties:
        input:
          oneOf:
            - type: string
              minLength: 1
              description: Non-empty user message. Whitespace-only values are invalid.
            - type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/AgentChatMessage'
    AgentChatMessage:
      type: object
      additionalProperties: false
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum:
            - user
            - assistant
            - developer
        content:
          type: string
          description: Message text. User messages must not be whitespace-only.
    CreatedAgent:
      type: object
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    AgentDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - author
        - role
      properties:
        id:
          type: string
        name:
          type: string
        author:
          type: string
          format: email
        role:
          type: string
          description: Role instructions for the agent. Empty string means no role instructions are set.
    AgentSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - statusCode
        - message
        - error
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
  responses:
    BadRequest:
      description: Request payload is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication token is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated user is not the author of the requested agent.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: No agent exists for the requested identifier.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ResponseGenerationFailed:
      description: The response-generation provider failed or is not configured.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'